Add default "posted on" text in a new partial template
If sites don't have .Site.Params.dateFormat set, then Hugo aborts
when trying to render {{ .Date.Format .Site.Params.dateFormat }},
leading to the index page missing post previews and the footer,
and post pages having all the content styled with the "post-meta"
class.

To fix this, specify a default value, and put it in a new partial
template so that there is only one place we have to change it if
we need to.