Fix figure shortcode with attrlink
There were a couple of problems when using the custom figure
shortcode with the attrlink parameter. First, there was a
scoping issue that caused the hugo templates to fail when
attempting to get the attr parameter. Second, the css for
the anchor tag placed the link in the wrong place.